Validate querystate model
https://coveord.atlassian.net/browse/JSUI-2586
