+++ src/ui/DynamicFacet/DynamicFacetValues/DynamicRangeFacetValueCreator.ts
@@ -2,23 +2,27 @@ import { DynamicRangeFacet } from '../DynamicRangeFacet';
 import { ValueCreator } from './DynamicFacetValues';
 import { IFacetResponseValue } from '../../../rest/Facet/FacetResponse';
 import { DynamicFacetValue } from './DynamicFacetValue';
-import { IRangeValue, RangeEndScope } from '../../../rest/RangeValue';
+import { IRangeValue } from '../../../rest/RangeValue';
 import { FacetValueState } from '../../../rest/Facet/FacetValueState';
 import { DynamicRangeFacetValueUtils as RangeUtils } from './DynamicRangeFacetValueUtils';
 
 export class DynamicRangeFacetValueCreator implements ValueCreator {
   constructor(private facet: DynamicRangeFacet) {}
 
+  private get valueFormat() {
+    return this.facet.options.valueFormat;
+  }
+
   public createFromRange(unvalidatedRange: IRangeValue, index: number) {
-    const range = RangeUtils.validateRange(unvalidatedRange, this.facet.options.valueFormat);
+    const range = RangeUtils.validateRange(unvalidatedRange, this.valueFormat);
     if (!range) {
-      this.facet.logger.error(`Unvalid range for ${this.facet.options.valueFormat} format`, unvalidatedRange);
+      this.facet.logger.error(`Unvalid range for ${this.valueFormat} format`, unvalidatedRange);
       return null;
     }
 
     const displayValue = range.label
       ? range.label
-      : RangeUtils.formatRangeValue(range, this.facet.options.valueFormat, this.facet.options.valueSeparator);
+      : RangeUtils.formatRangeValue(range, this.valueFormat, this.facet.options.valueSeparator);
 
     return new DynamicFacetValue(
       {
@@ -50,24 +54,10 @@ export class DynamicRangeFacetValueCreator implements ValueCreator {
     );
   }
 
-  private extractRangeFromValue(value: string): IRangeValue {
-    const valueRegex = new RegExp(`^(.+)\\.\\.(.+)(${RangeEndScope.Inclusive}|${RangeEndScope.Exclusive})$`);
-    const startAndEnd = valueRegex.exec(value);
-    if (!startAndEnd) {
-      this.facet.logger.error('Facet range value invalid', value);
-      return null;
-    }
-
-    return {
-      start: RangeUtils.validateRangeValue(startAndEnd[1], this.facet.options.valueFormat),
-      end: RangeUtils.validateRangeValue(startAndEnd[2], this.facet.options.valueFormat),
-      endInclusive: startAndEnd[3] === RangeEndScope.Inclusive
-    };
-  }
-
   public createFromValue(value: string) {
-    const range = this.extractRangeFromValue(value);
+    const range = RangeUtils.rangeFromValue(value, this.valueFormat);
     if (!range) {
+      this.facet.logger.error('Facet range value invalid', value);
       return null;
     }
